[NON-FUNCTIONAL] This is a simple logging plugin.
This is a channel-logging plugin. It produces a channel.log file for each
channel it is configured to reside in and logs events to it. This could
eventually be put up on a web page somewhere for our inane conversations to
horrify all political scientists.

The plugin is *mostly* written and un-tested. However, there is also the
early beginnings of a method for testing plugins. If you need to write to a
file during testing, you can give your plugin a file factory, retrieve the file
it wrote to as check that it's got the right things in it without ever going to
the file system.

This technique should be familiar to any Java programmer ;-)

The rest of the testing framework is quite shit. I really need to some how get
the test cases discoverable using `python -m unittest discover` but that can
wait. The most pressing concern right now is being able to coalesce the tests
into a single test/ directory and not have such import errors.

The logging module also currently fails to load, due to import errors about the
import of the file factory. It also does not configure itself.